Comment added

+++ models/productModel.go
@@ -11,7 +11,7 @@ type Product struct {
 	ProductName             string         `json:"product_name"`
 	ProductDescription      string         `json:"product_description"`
 	ProductImages           pq.StringArray `json:"product_images" gorm:"type:text[]"`
-	ProductPrice            float64        `json:"product_price"`
+	ProductPrice            uint32         `json:"product_price"`
 	CompressedProductImages pq.StringArray `json:"compressed_product_images" gorm:"type:text[]"`
 	CreatedAt               time.Time      `json:"created_at"`
 	UpdatedAt               time.Time      `json:"updated_at"`
