+++ models/userModel.go
@@ -8,8 +8,8 @@ type User struct {
 	ID        int       `gorm:"primary_key" json:"id"`
 	Name      string    `json:"name"`
 	Mobile    string    `json:"mobile"`
-	Latitude  float64   `json:"latitude"`
-	Longitude float64   `json:"longitude"`
+	Latitude  float32   `json:"latitude"`
+	Longitude float32   `json:"longitude"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
